минорные правки
edit-form-template.js
point.js

WIP
